requirements.txt
Django==1.10.6
django-model-utils==2.6.1
easy-thumbnails==2.3
django-bootstrap-pagination==1.6.2
MySQL-python==1.2.5
